+++ plugin/core/src/main/java/group/aelysium/rustyconnector/core/lib/algorithm/SortableInteger.java
@@ -10,7 +10,7 @@ public class SortableInteger implements ISortable {
     }
 
     @Override
-    public int sortIndex() {
+    public double sortIndex() {
         return this.value;
     }
